requirements.txt
# Core dependencies
python-dotenv>=1.0.0
langchain>=0.1.0
langgraph>=0.0.19

# LLM providers
openai>=1.3.0
google-generativeai>=0.3.0
langchain-google-genai>=0.0.5

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=3.2.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# Document processing
reportlab>=4.0.0
PyPDF2>=3.0.0
pillow>=10.0.0
python-docx>=1.1.0
openpyxl>=3.1.2

# Web interface
fastapi>=0.104.0
uvicorn>=0.23.0
python-multipart>=0.0.6
jinja2>=3.1.2

# Data handling
pandas>=2.1.3
numpy>=1.25.2

# Utilities
pydantic>=2.4.0
typing-extensions>=4.8.0
json5>=0.9.14
tqdm>=4.66.1
requests>=2.31.0
aiofiles>=23.2.1